Правки по комментариям наставника
